# Update text operator state to be resilient against exceptions in blocks