--- a/spec/spec_helper.rb
+++ b/spec/spec_helper.rb
@@ -10,6 +10,7 @@ end
 require_relative '../lib/pdf/core'
 
 require 'rspec'
+require 'securerandom'
 require 'pdf/reader'
 require 'pdf/inspector'