Before this PR, exceptions raised in the blocks wrapping text operator state changes could corrupt the document state.  If the operator state is being set to a value different than the current document state, and the block raises an exception, then both the in-memory variable and the in-PDF page state were corrupted.  This PR adds ensure blocks to guarantee that the value of both the in-memory and the PDF state is appropriately reset.

This PR also adds some basic specs for text_rendering_mode, as those specs were absent and this PR should test the error case for that operator state as well.

There is some additional opportunity to dry up the code, as the "real" operators (character_spacing, word_spacing, horizontal_text_scaling) are all essentially identical.