@gettalong 2022-02-15
I don't have the time to have a thorough look at it but there are now a lot of indirections there, e.g. `horizontal_scaling` -> `wrap_and_restore_horizontal_text_scaling` -> `update_horizontal_text_scaling_state` -> `update_real_text_state`. Since this is performance critical code, have you tested how it performs now? This may be the reason why there was the duplication in the first place. 

Also there will now be more string creations, e.g. due to `update_real_text_state`.