@petergoldstein 2022-02-15
@gettalong There's not a particular increase in the level of abstraction with this PR, simply a change in the organization and breakdown of the methods.  Aside from adding the ensure, removing the use of an unneeded local variable, and a slight change to string interpolation for the operators that take a real value, there's nothing more here than a bunch of extract methods for clarity.  The compiler will wind up inlining those in practice.

Regarding those specific items:

1. Adding the ensure is required to address the issue (it's the core of the problem).  It may wind up having a (very minor) performance hit, but it's not optional.
2. Replacing the use of a local variable with use of an instance variable in each of the `update_<operator>_state` should be cost-free
3. The change to string interpolation may (and I stress the may) have a minor cost, because we're replacing a single entry interpolation like this one:

`"\n#{PDF::Core.real(amount)} Tc"`

with

`"\n#{PDF::Core.real(amount)} #{operator}"`

where operator is a constant String.  I'm skeptical that this is going to be material in either allocations or time, but if it does we can change it back.

Where we WOULD introduce a higher level of abstraction is with further DRYing.  Essentially all of the operators that use unscaled text units (character_spacing, horizontal_text_scaling, word_spacing, and if we introduced leading and/or rise) have the exact same behavior and just set different instance variables and write different operator strings to the PDF.  Otherwise they're essentially identical.  So we could easily write a single abstract method that handles all those cases.  I held off from doing that precisely because I don't have a good benchmark.