@gettalong 2022-02-15
@petergoldstein I personally wouldn't probably introduce the `update_..._state` methods since it makes following the code - in these cases - needlessly complicated. But if they don't introduce much overhead, it is fine. Just maybe put them underneath their `wrap_and...` methods, then one doesn't have to jump around so much when reading the code.

I ran a small benchmark using benchmark-driver and the following benchmark file:

~~~
prelude: |
  x = 5.123

  def update(y)
    "\n#{y} Tw"
  end

benchmark:
  old: '"\n#{x} Tw"'
  new: 'update(x)'
~~~

Results:

~~~
Warming up --------------------------------------
                 old     1.246M i/s -      1.277M times in 1.025051s (802.47ns/i)
                 new   976.677k i/s -      1.023M times in 1.047560s (1.02μs/i)
Calculating -------------------------------------
                 old     1.322M i/s -      3.738M times in 2.827051s (756.21ns/i)
                 new     1.039M i/s -      2.930M times in 2.820546s (962.63ns/i)

Comparison:
                 old:   1322387.7 i/s
                 new:   1038816.8 i/s - 1.27x  slower
~~~

So using an intermediate method call makes this definitely slower. However, since this is a micro-benchmark, one would have to create a sample Prawn script to really test this out since other parts of Prawn will probably trump this change.