@petergoldstein 2022-02-15
@gettalong So I'm not going to get into the benchmarking - as you note, that's a microbenchmark.  Run in a standalone script, with a net impact of about 200 nsec per invocation.  Without any meaningful GC, etc.  If you want to run the HexaPDF benchmarks against the changes, we can take a look at that.

Regarding the DRY question - in my view there are basically two ways to DRY up this common logic.  One is by generating the methods at class definition time and the other is dynamically dispatching on invocation.  Depending on how you define meta programming one or both may qualify.

In either case we'd be introducing substantial abstraction, and there are pros and cons.  Having a map of the text state operators that basically reproduces the one in the PDF spec (Table 105 in the 1.7 spec), and generating or dispatching the methods from there might make the underlying logic clearer.  It also might make the code harder to read or less performant.  It would certainly make it easier to add new operators.  `pdf-reader` does something similar on the reading side, although obviously that's less performance sensitive

Given this PR is approved I'm going to merge it in and I can make overall adjustments when I put in the PR for the leading and rise operators, which should be shortly.